Pause & Game over
Addition of Pause and Game over screens to Level 1.
