+++ Assets/_Script/GameOverScreen.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class GameOverScreen : MonoBehaviour
+{
+    public void Retry()
+    {
+        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+    }
+    public void ReturnToTitle()
+    {
+        SceneManager.LoadScene(0);  //Loads the title screen (Scene 0)
+    }
+}
