+++ Assets/_Script/GameOverScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b66259f9ebeb0d4c980a81d93fb82bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
