+++ Assets/_Script/PauseScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0139a08aa9cc1dc4da002de29e81cb9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
