+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,15 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/TitleScreen.unity
+    guid: a5171cec4ba0e664e803106cb458f33e
+  - enabled: 1
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
+  - enabled: 1
+    path: Assets/Scenes/Level 1.unity
+    guid: ee6f9a89c816480409235b958867cf68
+  - enabled: 1
+    path: Assets/Scenes/Final Level.unity
+    guid: 3e71c80abf94ed84a9d3983e8f7d5214
   m_configObjects: {}
